Criando componente HomeMenu
Criando componente HomeMenu é transferindo código do componente Reserva para o HomeMenu
