+++ back/reserva-tech/src/App.jsx
@@ -1,14 +1,14 @@
 import { Routes, Route } from 'react-router-dom'
 import './App.css'
 import Home from './components/home'
-import Reserva from './components/Reserva'
+import HomeMenu from './components/HomeMenu'
 
 function App() {
   return (
     <>
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/reserva" element={<Reserva />} />
+        <Route path="/homemenu" element={<HomeMenu />} />
       </Routes>
     </>
   )
