+++ back/reserva-tech/src/components/HomeMenu.jsx
@@ -1,0 +1,44 @@
+import { useNavigate } from 'react-router-dom'
+import './HomeMenu.css'
+
+const HomeMenu = () => {
+    const navigate = useNavigate()
+
+    const handleVoltar = () => {
+        navigate('/')
+    }
+
+    return (
+        <>
+            <nav className="nav-cad">
+                <button onClick={handleVoltar}>Voltar</button>
+            </nav>
+            
+            <div className="centro">
+                <img src="/Imagens/Design sem nome (10).png" alt="ReservaTech Logo" />
+                
+                <div className="button">
+                    <div className="centro-cadastro">
+                        <p>Cadastrar</p>
+                        
+                        <div className="centro-cadastro-button">
+                            <a href="#">Cadastrar colaborador</a>
+                            <a href="#">Cadastrar equipamento</a>
+                        </div>
+                    </div>
+
+                    <div className="centro-reserva">
+                        <p>Reservas</p>
+                        
+                        <div className="centro-reserva-button">
+                            <a href="#">Reservar equipamento</a>
+                            <a href="#">Consultar reservas</a>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </>
+    )
+}
+
+export default HomeMenu
