--- back/reserva-tech/src/components/Reserva.css
@@ -1,154 +1,0 @@
-@charset "UTF-8";
-
-* {
-    margin: 0px;
-    padding: 0px;
-    box-sizing: border-box;
-}
-
-body {
-    width: 100%;
-    height: 100%;
-}
-
-.nav-cad {
-    background-color: gray;
-    padding: 20px;
-}
-
-.nav-cad button,
-.nav-cad a {
-    text-decoration: none;
-    color: white;
-    background-color: rgba(255, 255, 255, 0.26);
-    padding: 15px;
-    border-radius: 10px;
-    border: none;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-.nav-cad button:hover,
-.nav-cad a:hover {
-    background-color: rgb(0, 0, 0);
-}
-
-.centro {
-    background: linear-gradient(to right, #8056ff 50%, #148875 50%);
-    margin: 100px 50px;
-    height: 900px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.centro > img {
-    max-width: 60%;
-    margin: auto 20px;
-}
-
-.button {
-    margin-top: 100px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.button > a {
-    align-self: center;
-    justify-items: center;
-    border-radius: 7px;
-    background: white;
-    padding: 10px 20px;
-    margin-bottom: 10px;
-    text-decoration: none;
-    color: black;
-}
-
-.button > a:hover {
-    box-shadow: 3px 3px 10px #bfbbca;
-}
-
-.centro-cadastro {
-    margin-top: 50px;
-    margin-right: 300px;
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-}
-
-.centro-cadastro p {
-    color: white;
-    text-decoration: underline;
-}
-
-.centro-cadastro-button {
-    padding: 20px;
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-}
-
-.centro-cadastro-button a:nth-last-child(1) {
-    text-decoration: none;
-    color: black;
-    padding: 10px;
-    border-radius: 5px;
-    background-color: white;
-    margin: 5px;
-}
-
-.centro-cadastro-button a:nth-last-child(2) {
-    text-decoration: none;
-    color: black;
-    padding: 10px 13px;
-    border-radius: 5px;
-    background-color: white;
-    margin: 5px;
-}
-
-.centro-cadastro-button a:hover {
-    background-color: rgba(255, 255, 255, 0.63);
-}
-
-.centro-reserva {
-    margin-top: 50px;
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-}
-
-.centro-reserva p {
-    color: white;
-    text-decoration: underline;
-}
-
-.centro-reserva-button {
-    padding: 20px;
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-}
-
-.centro-reserva-button a:nth-last-child(1) {
-    text-decoration: none;
-    color: black;
-    padding: 10px 20px;
-    border-radius: 5px;
-    background-color: white;
-    margin: 5px;
-}
-
-.centro-reserva-button a:nth-last-child(2) {
-    text-decoration: none;
-    color: black;
-    padding: 10px 8px;
-    border-radius: 5px;
-    background-color: white;
-    margin: 5px;
-}
-
-.centro-reserva-button a:hover {
-    background-color: rgba(255, 255, 255, 0.63);
-}
